mv fixtures to fixtures
